refactor(cmd): split command files under 300 line limit
