+++ cmd/buildtemplates/main.go
@@ -1,8 +1,7 @@
 // Command buildtemplates creates base sandbox snapshots with Bridge pre-installed.
 //
 // It builds templates of different sizes (small, medium, large, xlarge) and
-// flavors (bridge, dev-box) on the configured sandbox provider. Currently
-// supports Daytona; designed to be extended for future providers.
+// flavors (bridge, dev-box) on the configured sandbox provider.
 //
 // Usage:
 //
@@ -21,310 +20,7 @@ import (
 	"os"
 	"strings"
 	"time"
-
-	daytona "github.com/daytonaio/daytona/libs/sdk-go/pkg/daytona"
-	"github.com/daytonaio/daytona/libs/sdk-go/pkg/types"
-	"github.com/usehiveloop/hiveloop/internal/model"
 )
-
-const (
-	baseImage         = "ubuntu:24.04"
-	bridgeDir         = "/usr/local/bin"
-	daytonaHome       = "/home/daytona"
-	bridgeReleasesURL = "https://github.com/usehiveloop/bridge/releases/download"
-
-	flavorBridge = "bridge"
-	flavorDevBox = "dev-box"
-)
-
-var basePackages = []string{
-	"curl",
-	"ca-certificates",
-	"git",
-	"jq",
-	"unzip",
-	"wget",
-	"openssh-client",
-}
-
-const nvmVersion = "v0.40.4"
-
-const goVersion = "1.24.2"
-
-const astGrepVersion = "0.42.1"
-
-// devToolPackages are CLI tools and server binaries that ship dormant in the
-// dev-box image. None of these start daemons at boot — the entrypoint only
-// runs Bridge at boot. Agents start postgres/redis explicitly.
-var devToolPackages = []string{
-	"build-essential",
-	"python3-pip",
-	"python3-venv",
-	"sqlite3",
-	"libsqlite3-dev",
-	"postgresql",
-	"postgresql-client",
-	"redis-server",
-	"ffmpeg",
-	"tmux",
-	"screen",
-	"zip",
-	"tar",
-	"gzip",
-	"xz-utils",
-	"zstd",
-	"bzip2",
-	"dnsutils",
-	"net-tools",
-	"httpie",
-	"openssl",
-	"nano",
-	"libxml2-utils",
-	"xmlstarlet",
-	"s3cmd",
-	"ripgrep",
-}
-
-// sizes re-exports model.TemplateSizes for local convenience.
-var sizes = model.TemplateSizes
-
-func bridgeDownloadURL(version string) string {
-	return fmt.Sprintf("%s/v%s/bridge-v%s-x86_64-unknown-linux-gnu.tar.gz",
-		bridgeReleasesURL, version, version)
-}
-
-// buildBaseImage installs the shared runtime layer used by every flavor:
-// system packages and the Bridge binary. Each flavor is responsible
-// for its own GitHub tooling, additional layers, workdir, and entrypoint.
-func buildBaseImage(bridgeVersion string) *daytona.DockerImage {
-	downloadURL := bridgeDownloadURL(bridgeVersion)
-
-	image := daytona.Base(baseImage)
-
-	// Switch to Hetzner's mirror — Ubuntu's archive throttles Hetzner IPs to ~120KB/s
-	// while Hetzner's own mirror serves at 6+ MB/s.
-	// archive.ubuntu.com/ubuntu → mirror.hetzner.de/ubuntu/packages
-	// security.ubuntu.com/ubuntu → mirror.hetzner.de/ubuntu/packages
-	image = image.Run(
-		`sed -i 's|http://archive.ubuntu.com/ubuntu|http://mirror.hetzner.de/ubuntu/packages|g; s|http://security.ubuntu.com/ubuntu|http://mirror.hetzner.de/ubuntu/packages|g' /etc/apt/sources.list.d/ubuntu.sources || ` +
-			`sed -i 's|http://archive.ubuntu.com/ubuntu|http://mirror.hetzner.de/ubuntu/packages|g; s|http://security.ubuntu.com/ubuntu|http://mirror.hetzner.de/ubuntu/packages|g' /etc/apt/sources.list`)
-
-	// System packages
-	image = image.AptGet(basePackages)
-
-	// Home directory + Bridge storage directory
-	image = image.Run(fmt.Sprintf("mkdir -p %s/.bridge", daytonaHome))
-
-	// Bridge binary
-	image = image.Run(fmt.Sprintf(
-		`curl -fsSL "%s" | tar -xzf - -C %s && chmod +x %s/bridge`,
-		downloadURL, bridgeDir, bridgeDir,
-	))
-
-	return image
-}
-
-// buildBridgeImage produces the default flavor: just the base runtime.
-func buildBridgeImage(bridgeVersion string) *daytona.DockerImage {
-	image := buildBaseImage(bridgeVersion)
-
-	image = image.Workdir(daytonaHome)
-	image = image.Entrypoint([]string{"/bin/sh", "-c", "mkdir -p /home/daytona/.bridge && /usr/local/bin/bridge >> /tmp/bridge.log 2>&1"})
-
-	return image
-}
-
-// buildDevBoxImage layers a developer toolchain on top of the base bridge
-// runtime: Node.js (via nvm), Chrome for Testing, and agent-browser
-// for AI-driven browser automation. agent-browser's daemon starts lazily
-// on the first CLI command — no pre-warming needed in the entrypoint.
-func buildDevBoxImage(bridgeVersion string) *daytona.DockerImage {
-	image := buildBaseImage(bridgeVersion)
-
-	// Install nvm + Node LTS into a system-wide location and symlink the
-	// resulting binaries into /usr/local/bin so non-login shells (and Bridge)
-	// pick them up without needing to source nvm.sh.
-	nvmInstall := strings.Join([]string{
-		"set -eux",
-		"export NVM_DIR=/usr/local/nvm",
-		"mkdir -p $NVM_DIR",
-		"curl -fsSL https://raw.githubusercontent.com/nvm-sh/nvm/" + nvmVersion + "/install.sh | bash",
-		". $NVM_DIR/nvm.sh",
-		"nvm install --lts",
-		"NODE_BIN=$(nvm which current)",
-		"NODE_DIR=$(dirname $NODE_BIN)",
-		"ln -sf $NODE_BIN /usr/local/bin/node",
-		"ln -sf $NODE_DIR/npm /usr/local/bin/npm",
-		"ln -sf $NODE_DIR/npx /usr/local/bin/npx",
-	}, " && ")
-	image = image.Run("bash -c '" + nvmInstall + "'")
-
-	// Install agent-browser and other global CLIs.
-	// --prefix=/usr/local forces npm to drop the bin shims into /usr/local/bin
-	// instead of nvm's per-version prefix (/usr/local/nvm/versions/node/<v>/bin),
-	// which is NOT on the default PATH that the Bridge entrypoint sees.
-	image = image.Run("npm install -g --prefix=/usr/local agent-browser")
-
-	// GitHub CLI (official apt repository).
-	image = image.Run(
-		"mkdir -p -m 755 /etc/apt/keyrings && " +
-			"wget -qO- https://cli.github.com/packages/githubcli-archive-keyring.gpg | tee /etc/apt/keyrings/githubcli-archive-keyring.gpg > /dev/null && " +
-			"chmod go+r /etc/apt/keyrings/githubcli-archive-keyring.gpg && " +
-			"echo 'deb [arch=amd64 signed-by=/etc/apt/keyrings/githubcli-archive-keyring.gpg] https://cli.github.com/packages stable main' > /etc/apt/sources.list.d/github-cli.list && " +
-			"apt-get update && apt-get install -y gh && rm -rf /var/lib/apt/lists/*")
-
-	// Download Chrome for Testing and install its Linux shared-library
-	// dependencies (libnss3, libatk, libgbm, fonts, etc.) in one step.
-	// agent-browser auto-detects container environments and adds --no-sandbox.
-	// Sandbox runs as root, so Chrome installs to /root/.agent-browser/.
-	image = image.Run("agent-browser install --with-deps")
-
-	// Dev tools: compilers, databases, media, terminal multiplexers,
-	// network diagnostics, archive utilities, editors.
-	// All are dormant binaries — no daemons start at boot.
-	image = image.AptGet(devToolPackages)
-
-	// yq (YAML/JSON/TOML processor) — not in apt, installed as a standalone binary.
-	image = image.Run(
-		`curl -fsSL https://github.com/mikefarah/yq/releases/latest/download/yq_linux_amd64 -o /usr/local/bin/yq && chmod +x /usr/local/bin/yq`,
-	)
-
-	// ast-grep (structural code search/rewrite) — not in apt, installed from
-	// the official GitHub release. Ships two binaries: ast-grep and sg.
-	image = image.Run(fmt.Sprintf(
-		`curl -fsSL https://github.com/ast-grep/ast-grep/releases/download/%s/app-x86_64-unknown-linux-gnu.zip -o /tmp/ast-grep.zip && `+
-			`unzip -o /tmp/ast-grep.zip -d /usr/local/bin/ && `+
-			`chmod +x /usr/local/bin/ast-grep /usr/local/bin/sg && `+
-			`rm /tmp/ast-grep.zip`,
-		astGrepVersion,
-	))
-
-	// Go (official tarball, symlinked into /usr/local/bin for default PATH).
-	image = image.Run(fmt.Sprintf(
-		"curl -fsSL https://go.dev/dl/go%s.linux-amd64.tar.gz | tar -C /usr/local -xzf - && "+
-			"ln -sf /usr/local/go/bin/go /usr/local/bin/go && "+
-			"ln -sf /usr/local/go/bin/gofmt /usr/local/bin/gofmt",
-		goVersion,
-	))
-
-	// Rust via rustup (the standard version manager). RUSTUP_HOME and
-	// CARGO_HOME are set to system-wide paths; key binaries are symlinked
-	// into /usr/local/bin so non-login shells find them.
-	image = image.Env("RUSTUP_HOME", "/usr/local/rustup")
-	image = image.Env("CARGO_HOME", "/usr/local/cargo")
-	image = image.Run(
-		"curl --proto '=https' --tlsv1.2 -sSf https://sh.rustup.rs | sh -s -- -y --no-modify-path && " +
-			"ln -sf /usr/local/cargo/bin/rustc /usr/local/bin/rustc && " +
-			"ln -sf /usr/local/cargo/bin/cargo /usr/local/bin/cargo && " +
-			"ln -sf /usr/local/cargo/bin/rustup /usr/local/bin/rustup",
-	)
-
-	// Pre-install all language servers supported by Bridge so agents don't
-	// pay cold-start cost on first LSP request.
-	image = image.Run("/usr/local/bin/bridge install-lsp all")
-
-	// Git credential helper — fetches GitHub tokens from the control plane
-	// on demand. Token never touches disk; git calls this on every auth request.
-	image = image.Run(
-		`printf '#!/bin/sh\ncurl -sf -X POST -H "Authorization: Bearer $BRIDGE_CONTROL_PLANE_API_KEY" "$HIVELOOP_GIT_CREDENTIALS_URL"\n' > /usr/local/bin/git-credential-hiveloop && ` +
-			`chmod +x /usr/local/bin/git-credential-hiveloop`,
-	)
-	image = image.Run("git config --system credential.helper /usr/local/bin/git-credential-hiveloop")
-	image = image.Run("git config --system user.name hiveloop")
-	image = image.Run("git config --system user.email help@hiveloop.com")
-	image = image.Run("git config --global user.name hiveloop")
-	image = image.Run("git config --global user.email help@hiveloop.com")
-
-	// gh CLI wrapper — fetches a fresh GitHub token on every invocation so
-	// `gh issue create`, `gh issue list`, etc. work without manual auth.
-	// /usr/local/bin takes precedence over /usr/bin on PATH.
-	image = image.Run(
-		`printf '#!/bin/sh\nexport GH_NO_KEYRING=1\nexport GH_TOKEN=$(curl -sf -X POST -H "Authorization: Bearer $BRIDGE_CONTROL_PLANE_API_KEY" "$HIVELOOP_GIT_CREDENTIALS_URL" | grep password | cut -d= -f2)\nexec /usr/bin/gh "$@"\n' > /usr/local/bin/gh-wrapper && ` +
-			`chmod +x /usr/local/bin/gh-wrapper && ` +
-			`ln -sf /usr/local/bin/gh-wrapper /usr/local/bin/gh`,
-	)
-
-	image = image.Workdir(daytonaHome)
-	// agent-browser's daemon starts lazily on the first CLI command, so
-	// the entrypoint only needs to launch Bridge. No daemon pre-warming
-	// or health check is required.
-	image = image.Entrypoint([]string{"/bin/sh", "-c",
-		"mkdir -p /home/daytona/.bridge && " +
-			"exec /usr/local/bin/bridge >> /tmp/bridge.log 2>&1"})
-
-	return image
-}
-
-// snapshotName returns the published snapshot name for a flavor + version + size.
-// The dev-box flavor uses the hiveloop-dev-box-<size>-v<version> scheme; the
-// default bridge flavor keeps its historical name so existing snapshots aren't
-// orphaned.
-func snapshotName(flavor, bridgeVersion, size string) string {
-	switch flavor {
-	case flavorDevBox:
-		return fmt.Sprintf("hiveloop-dev-box-%s-v%s", size, bridgeVersion)
-	default:
-		return fmt.Sprintf("hiveloop-bridge-%s-%s", strings.ReplaceAll(bridgeVersion, ".", "-"), size)
-	}
-}
-
-func buildDaytona(ctx context.Context, flavor, bridgeVersion string, targetSizes []string) error {
-	client, err := daytona.NewClientWithConfig(&types.DaytonaConfig{
-		APIKey: os.Getenv("SANDBOX_PROVIDER_KEY"),
-		APIUrl: os.Getenv("SANDBOX_PROVIDER_URL"),
-		Target: os.Getenv("SANDBOX_TARGET"),
-	})
-	if err != nil {
-		return fmt.Errorf("creating daytona client: %w", err)
-	}
-	defer client.Close(ctx)
-
-	var image *daytona.DockerImage
-	switch flavor {
-	case flavorBridge:
-		image = buildBridgeImage(bridgeVersion)
-	case flavorDevBox:
-		image = buildDevBoxImage(bridgeVersion)
-	default:
-		return fmt.Errorf("unknown flavor: %s (valid: %s, %s)", flavor, flavorBridge, flavorDevBox)
-	}
-	log.Printf("Generated Dockerfile (flavor=%s):\n%s\n", flavor, image.Dockerfile())
-
-	for _, sizeName := range targetSizes {
-		size, ok := sizes[sizeName]
-		if !ok {
-			return fmt.Errorf("unknown size: %s", sizeName)
-		}
-
-		name := snapshotName(flavor, bridgeVersion, size.Name)
-		log.Printf("Building snapshot %q (cpu=%d, mem=%dGB, disk=%dGB)...",
-			name, size.CPU, size.Memory, size.Disk)
-
-		resources := &types.Resources{
-			CPU:    size.CPU,
-			Memory: size.Memory,
-			Disk:   size.Disk,
-		}
-
-		snapshot, logChan, err := client.Snapshot.Create(ctx, &types.CreateSnapshotParams{
-			Name:      name,
-			Image:     image,
-			Resources: resources,
-		})
-		if err != nil {
-			return fmt.Errorf("creating snapshot %q: %w", name, err)
-		}
-
-		// Stream build logs
-		for line := range logChan {
-			log.Printf("[%s] %s", name, line)
-		}
-
-		log.Printf("Snapshot %q created successfully (id=%s)", name, snapshot.Name)
-	}
-
-	return nil
-}
 
 func main() {
 	version := flag.String("version", "", "Bridge version to install (required, e.g. 0.10.0)")
@@ -346,7 +42,6 @@ func main() {
 		os.Exit(1)
 	}
 
-	// Determine which sizes to build
 	var targetSizes []string
 	if *size == "all" {
 		targetSizes = []string{"small", "medium", "large", "xlarge"}
