+++ cmd/fetchactions-oas2/parse.go
@@ -10,59 +10,6 @@ import (
 	v2high "github.com/pb33f/libopenapi/datamodel/high/v2"
 )
 
-// ActionDef mirrors the catalog ActionDef for JSON output.
-type ActionDef struct {
-	DisplayName    string           `json:"display_name"`
-	Description    string           `json:"description"`
-	Access         string           `json:"access"`
-	ResourceType   string           `json:"resource_type"`
-	Parameters     json.RawMessage  `json:"parameters"`
-	Execution      *ExecutionConfig `json:"execution,omitempty"`
-	ResponseSchema string           `json:"response_schema,omitempty"` // ref into top-level schemas map
-}
-
-// ExecutionConfig mirrors the catalog ExecutionConfig.
-type ExecutionConfig struct {
-	Method       string            `json:"method"`
-	Path         string            `json:"path"`
-	BodyMapping  map[string]string `json:"body_mapping,omitempty"`
-	QueryMapping map[string]string `json:"query_mapping,omitempty"`
-	Headers      map[string]string `json:"headers,omitempty"`
-	ResponsePath string            `json:"response_path,omitempty"`
-}
-
-// jsonSchemaProperty represents a single property in the JSON Schema.
-type jsonSchemaProperty struct {
-	Type        string `json:"type"`
-	Description string `json:"description,omitempty"`
-}
-
-// SchemaProperty describes a single property in a flattened response schema.
-type SchemaProperty struct {
-	Type        string `json:"type"`
-	Description string `json:"description,omitempty"`
-	Nullable    bool   `json:"nullable,omitempty"`
-	SchemaRef   string `json:"schema_ref,omitempty"` // references another schema for nested object resolution
-}
-
-// FlatSchema is a flattened top-level-only representation of a response schema.
-type FlatSchema struct {
-	Type       string                    `json:"type"`
-	Properties map[string]SchemaProperty `json:"properties,omitempty"`
-	Items      *FlatSchemaRef            `json:"items,omitempty"` // for array types
-}
-
-// FlatSchemaRef references another schema by name (for array item types).
-type FlatSchemaRef struct {
-	Ref string `json:"$ref,omitempty"`
-}
-
-// ParseResult holds parsed actions and the referenced response schemas.
-type ParseResult struct {
-	Actions map[string]ActionDef
-	Schemas map[string]FlatSchema
-}
-
 // parseSpec parses a Swagger 2.0 spec and returns actions + referenced response schemas.
 func parseSpec(specData []byte, cfg ServiceConfig) (*ParseResult, error) {
 	doc, err := libopenapi.NewDocument(specData)
@@ -87,7 +34,6 @@ func parseSpec(specData []byte, cfg ServiceConfig) (*ParseResult, error) {
 		return &ParseResult{Actions: actions, Schemas: schemas}, nil
 	}
 
-	// Build definitions lookup for resolving $ref in response schemas.
 	definitions := model.Model.Definitions
 	definitionsMap := make(map[string]*highbase.SchemaProxy)
 	if definitions != nil && definitions.Definitions != nil {
@@ -170,417 +116,4 @@ func getV2Operations(pi *v2high.PathItem) map[string]*v2high.Operation {
 		ops["DELETE"] = pi.Delete
 	}
 	return ops
-}
-
-// matchesFilters checks if a path matches include/exclude filters.
-func matchesFilters(path string, includes, excludes []string) bool {
-	if len(includes) > 0 {
-		matched := false
-		for _, prefix := range includes {
-			if strings.HasPrefix(path, prefix) {
-				matched = true
-				break
-			}
-		}
-		if !matched {
-			return false
-		}
-	}
-	for _, prefix := range excludes {
-		if strings.HasPrefix(path, prefix) {
-			return false
-		}
-	}
-	return true
-}
-
-// deriveV2ActionKey produces the action key and display name from a Swagger operation.
-func deriveV2ActionKey(op *v2high.Operation, method, path string) (string, string) {
-	var key string
-	if op.OperationId != "" {
-		key = toSnakeCase(op.OperationId)
-	} else {
-		key = fallbackActionKey(method, path)
-	}
-
-	displayName := ""
-	if op.Summary != "" {
-		displayName = op.Summary
-	} else {
-		displayName = toDisplayName(key)
-	}
-	if len(displayName) > 80 {
-		displayName = displayName[:77] + "..."
-	}
-	return key, displayName
-}
-
-// inferV2ResourceType uses tags to determine resource type.
-func inferV2ResourceType(op *v2high.Operation, tagMap map[string]string) string {
-	if tagMap == nil {
-		return ""
-	}
-	for _, tag := range op.Tags {
-		if rt, ok := tagMap[strings.ToLower(tag)]; ok {
-			return rt
-		}
-		if rt, ok := tagMap[tag]; ok {
-			return rt
-		}
-	}
-	return ""
-}
-
-// buildV2ParamsAndExecution extracts Swagger 2.0 parameters and builds execution config.
-func buildV2ParamsAndExecution(op *v2high.Operation, pathItemParams []*v2high.Parameter, method, path string, extraHeaders map[string]string) (json.RawMessage, *ExecutionConfig) {
-	properties := make(map[string]jsonSchemaProperty)
-	var required []string
-	queryMapping := make(map[string]string)
-	bodyMapping := make(map[string]string)
-
-	processParam := func(param *v2high.Parameter) {
-		name := param.Name
-		if name == "" {
-			return
-		}
-		// Skip formData (file uploads).
-		if param.In == "formData" {
-			return
-		}
-		if _, exists := properties[name]; exists {
-			return
-		}
-
-		switch param.In {
-		case "body":
-			// Body param in Swagger 2.0: extract top-level properties from schema.
-			if param.Schema != nil {
-				s := param.Schema.Schema()
-				if s != nil && s.Properties != nil {
-					for prop := s.Properties.First(); prop != nil; prop = prop.Next() {
-						propName := prop.Key()
-						propSchema := prop.Value()
-						if _, exists := properties[propName]; exists {
-							continue
-						}
-						propType := "string"
-						propDesc := ""
-						if propSchema != nil {
-							ps := propSchema.Schema()
-							if ps != nil {
-								if len(ps.Type) > 0 {
-									propType = ps.Type[0]
-								}
-								propDesc = ps.Description
-								if len(propDesc) > 200 {
-									propDesc = propDesc[:197] + "..."
-								}
-							}
-						}
-						properties[propName] = jsonSchemaProperty{
-							Type:        propType,
-							Description: propDesc,
-						}
-						bodyMapping[propName] = propName
-					}
-					// Check required from body schema.
-					for _, r := range s.Required {
-						if _, exists := properties[r]; exists {
-							required = append(required, r)
-						}
-					}
-				}
-			}
-		case "path", "query":
-			paramType := "string"
-			if param.Type != "" {
-				paramType = param.Type
-			}
-			desc := param.Description
-			if len(desc) > 200 {
-				desc = desc[:197] + "..."
-			}
-			properties[name] = jsonSchemaProperty{
-				Type:        paramType,
-				Description: desc,
-			}
-			if param.In == "path" {
-				required = append(required, name)
-			} else {
-				queryMapping[name] = name
-			}
-		}
-	}
-
-	// Operation-level params first, then path-item-level.
-	for _, param := range op.Parameters {
-		processParam(param)
-	}
-	for _, param := range pathItemParams {
-		processParam(param)
-	}
-
-	// Fallback: extract path template params.
-	for _, match := range pathParamRe.FindAllStringSubmatch(path, -1) {
-		paramName := match[1]
-		if _, exists := properties[paramName]; !exists {
-			properties[paramName] = jsonSchemaProperty{
-				Type:        "string",
-				Description: paramName,
-			}
-			required = append(required, paramName)
-		}
-	}
-
-	schema := map[string]any{
-		"type":       "object",
-		"properties": properties,
-	}
-	if len(required) > 0 {
-		schema["required"] = dedupStrings(required)
-	}
-	paramsJSON, _ := json.Marshal(schema)
-
-	exec := &ExecutionConfig{
-		Method: method,
-		Path:   path,
-	}
-	if len(queryMapping) > 0 {
-		exec.QueryMapping = queryMapping
-	}
-	if len(bodyMapping) > 0 {
-		exec.BodyMapping = bodyMapping
-	}
-	if len(extraHeaders) > 0 {
-		exec.Headers = extraHeaders
-	}
-
-	return paramsJSON, exec
-}
-
-// dedupStrings removes duplicates from a string slice.
-// readHintPrefixes indicate a read even when HTTP method is POST.
-var readHintPrefixes = []string{
-	"list", "get", "search", "find", "query", "read", "fetch", "check",
-	"show", "describe", "lookup", "retrieve", "export", "download", "view",
-}
-
-var readHintPathSuffixes = []string{
-	"/search", "/query", "/filter", "/find", "/list", "/lookup", "/export",
-}
-
-// inferAccess determines "read" or "write" from method, action key, and path.
-func inferAccess(method, actionKey, path string) string {
-	switch method {
-	case "GET":
-		return "read"
-	case "POST":
-		keyLower := strings.ToLower(actionKey)
-		for _, prefix := range readHintPrefixes {
-			if strings.Contains(keyLower, prefix) {
-				return "read"
-			}
-		}
-		pathLower := strings.ToLower(path)
-		for _, suffix := range readHintPathSuffixes {
-			if strings.HasSuffix(pathLower, suffix) {
-				return "read"
-			}
-		}
-		return "write"
-	default:
-		return "write"
-	}
-}
-
-func dedupStrings(ss []string) []string {
-	seen := make(map[string]bool, len(ss))
-	result := make([]string, 0, len(ss))
-	for _, s := range ss {
-		if !seen[s] {
-			seen[s] = true
-			result = append(result, s)
-		}
-	}
-	return result
-}
-
-// extractV2ResponseSchema looks at the 200 response of a Swagger 2.0 operation,
-// resolves the schema ref name, flattens it into the schemas map, and returns
-// the ref name. Returns "" if no usable response schema is found.
-func extractV2ResponseSchema(op *v2high.Operation, definitions map[string]*highbase.SchemaProxy, schemas map[string]FlatSchema) string {
-	if op.Responses == nil || op.Responses.Codes == nil {
-		return ""
-	}
-
-	for _, code := range []string{"200", "201"} {
-		resp := op.Responses.Codes.GetOrZero(code)
-		if resp == nil || resp.Schema == nil {
-			continue
-		}
-
-		schema := resp.Schema.Schema()
-		if schema == nil {
-			continue
-		}
-
-		// Case 1: Direct $ref to a definition.
-		refName := extractRefName(resp.Schema)
-		if refName != "" {
-			flattenV2Definition(refName, definitions, schemas)
-			return refName
-		}
-
-		// Case 2: Array with $ref items.
-		if len(schema.Type) > 0 && schema.Type[0] == "array" && schema.Items != nil && schema.Items.A != nil {
-			itemRef := extractRefName(schema.Items.A)
-			if itemRef != "" {
-				// Store the array wrapper schema pointing to the item schema.
-				arraySchemaName := itemRef + "_list"
-				if _, exists := schemas[arraySchemaName]; !exists {
-					flattenV2Definition(itemRef, definitions, schemas)
-					schemas[arraySchemaName] = FlatSchema{
-						Type:  "array",
-						Items: &FlatSchemaRef{Ref: itemRef},
-					}
-				}
-				return arraySchemaName
-			}
-		}
-
-		// Case 3: Inline object — flatten it directly using the operation ID as name.
-		if len(schema.Type) > 0 && schema.Type[0] == "object" && schema.Properties != nil {
-			inlineName := toSnakeCase(op.OperationId) + "_response"
-			if _, exists := schemas[inlineName]; !exists {
-				flat := FlatSchema{
-					Type:       "object",
-					Properties: make(map[string]SchemaProperty),
-				}
-				for prop := schema.Properties.First(); prop != nil; prop = prop.Next() {
-					propName := prop.Key()
-					propProxy := prop.Value()
-					flat.Properties[propName] = flattenV2Property(propProxy)
-				}
-				schemas[inlineName] = flat
-			}
-			return inlineName
-		}
-	}
-
-	return ""
-}
-
-// extractRefName pulls the definition name from a schema proxy's $ref if present.
-func extractRefName(proxy *highbase.SchemaProxy) string {
-	if proxy == nil {
-		return ""
-	}
-	// GetReference returns the raw $ref string like "#/definitions/objs_message".
-	ref := proxy.GetReference()
-	if ref == "" {
-		return ""
-	}
-	const prefix = "#/definitions/"
-	if strings.HasPrefix(ref, prefix) {
-		return strings.TrimPrefix(ref, prefix)
-	}
-	// Fall back to last segment for any ref format.
-	parts := strings.Split(ref, "/")
-	return parts[len(parts)-1]
-}
-
-// flattenV2Definition resolves a Swagger 2.0 definition by name and stores its
-// flattened top-level properties in the schemas map. Nested $refs are collapsed
-// to their type only (no recursion).
-func flattenV2Definition(name string, definitions map[string]*highbase.SchemaProxy, schemas map[string]FlatSchema) {
-	if _, exists := schemas[name]; exists {
-		return
-	}
-
-	proxy, ok := definitions[name]
-	if !ok {
-		return
-	}
-	schema := proxy.Schema()
-	if schema == nil {
-		return
-	}
-
-	flat := FlatSchema{
-		Type:       "object",
-		Properties: make(map[string]SchemaProperty),
-	}
-
-	if len(schema.Type) > 0 {
-		flat.Type = schema.Type[0]
-	}
-
-	// For array types, try to resolve item ref.
-	if flat.Type == "array" && schema.Items != nil && schema.Items.A != nil {
-		itemRef := extractRefName(schema.Items.A)
-		if itemRef != "" {
-			flattenV2Definition(itemRef, definitions, schemas)
-			flat.Items = &FlatSchemaRef{Ref: itemRef}
-		}
-		schemas[name] = flat
-		return
-	}
-
-	if schema.Properties != nil {
-		for prop := schema.Properties.First(); prop != nil; prop = prop.Next() {
-			propName := prop.Key()
-			propProxy := prop.Value()
-			flat.Properties[propName] = flattenV2Property(propProxy)
-		}
-	}
-
-	schemas[name] = flat
-
-	// Transitively flatten any schemas referenced by properties via schema_ref.
-	for _, propDef := range flat.Properties {
-		if propDef.SchemaRef != "" {
-			flattenV2Definition(propDef.SchemaRef, definitions, schemas)
-		}
-	}
-}
-
-// flattenV2Property extracts type + description from a schema proxy without recursing.
-// When the property is a $ref to a named definition, the ref name is preserved
-// as SchemaRef so the frontend can resolve nested object types.
-func flattenV2Property(proxy *highbase.SchemaProxy) SchemaProperty {
-	prop := SchemaProperty{Type: "string"}
-	if proxy == nil {
-		return prop
-	}
-
-	refName := extractRefName(proxy)
-
-	schema := proxy.Schema()
-	if schema == nil {
-		if refName != "" {
-			prop.Type = "object"
-			prop.SchemaRef = refName
-		}
-		return prop
-	}
-
-	if len(schema.Type) > 0 {
-		prop.Type = schema.Type[0]
-	} else if refName != "" {
-		prop.Type = "object"
-	}
-
-	if refName != "" && prop.Type == "object" {
-		prop.SchemaRef = refName
-	}
-
-	if schema.Description != "" {
-		prop.Description = schema.Description
-	}
-
-	if schema.Nullable != nil && *schema.Nullable {
-		prop.Nullable = true
-	}
-
-	return prop
 }
